Update Dao_Base_Spot.php
Sanity checks

+++ lib/dao/Base/Dao_Base_Spot.php
@@ -270,7 +270,14 @@ class Dao_Base_Spot implements Dao_Spot
     public function updateSpotRating($spotMsgIdList)
     {
         // Empty list provided? Exit
-        if (count($spotMsgIdList) == 0) {
+        if (!is_array($spotMsgIdList) || count($spotMsgIdList) == 0) {
+            return;
+        } // if
+
+        // prepare a list of IN values
+        $msgIdList = $this->_conn->arrayKeyToIn($spotMsgIdList);
+ 
+        if (!isset($msgIdList) || $msgIdList == '') {
             return;
         } // if
 
@@ -285,7 +292,7 @@ class Dao_Base_Spot implements Dao_Spot
 										spots.messageid = commentsxover.nntpref 
 										AND spotrating BETWEEN 1 AND 10
 									 GROUP BY nntpref)
-							WHERE spots.messageid IN ('.$this->_conn->arrayKeyToIn($spotMsgIdList).')
+							WHERE spots.messageid IN ('.$msgIdList.')
 						');
         SpotTiming::stop(__CLASS__.'::'.__FUNCTION__, [$spotMsgIdList]);
     }
@@ -298,7 +305,14 @@ class Dao_Base_Spot implements Dao_Spot
     public function updateSpotCommentCount($spotMsgIdList)
     {
         // Empty list provided? Exit
-        if (count($spotMsgIdList) == 0) {
+        if (!is_array($spotMsgIdList) || count($spotMsgIdList) == 0) {
+            return;
+        } // if
+
+        // prepare a list of IN values
+        $msgIdList = $this->_conn->arrayKeyToIn($spotMsgIdList);
+ 
+        if (!isset($msgIdList) || $msgIdList == '') {
             return;
         } // if
 
@@ -310,7 +324,7 @@ class Dao_Base_Spot implements Dao_Spot
 									 WHERE 
 										spots.messageid = commentsxover.nntpref 
 									 GROUP BY nntpref)
-							WHERE spots.messageid IN ('.$this->_conn->arrayKeyToIn($spotMsgIdList).')
+							WHERE spots.messageid IN ('.$msgIdList.')
 						');
         SpotTiming::stop(__CLASS__.'::'.__FUNCTION__, [$spotMsgIdList]);
     }
@@ -323,7 +337,14 @@ class Dao_Base_Spot implements Dao_Spot
     public function updateSpotReportCount($spotMsgIdList)
     {
         // Empty list provided? Exit
-        if (count($spotMsgIdList) == 0) {
+        if (!is_array($spotMsgIdList) || count($spotMsgIdList) == 0) {
+            return;
+        } // if
+
+        // prepare a list of IN values
+        $msgIdList = $this->_conn->arrayKeyToIn($spotMsgIdList);
+ 
+        if (!isset($msgIdList) || $msgIdList == '') {
             return;
         } // if
 
@@ -335,7 +356,7 @@ class Dao_Base_Spot implements Dao_Spot
 									 WHERE 
 										spots.messageid = reportsxover.nntpref 
 									 GROUP BY nntpref)
-							WHERE spots.messageid IN ('.$this->_conn->arrayKeyToIn($spotMsgIdList).')
+							WHERE spots.messageid IN ('.$msgIdList.')
 						');
         SpotTiming::stop(__CLASS__.'::'.__FUNCTION__, [$spotMsgIdList]);
     }
@@ -347,15 +368,22 @@ class Dao_Base_Spot implements Dao_Spot
      */
     public function getDisposedSpots($spotMsgIdList)
     {
-        SpotTiming::start(__CLASS__.'::'.__FUNCTION__);
         $tmparray = [];
+
         // Empty list provided? Exit
-        if (count($spotMsgIdList) == 0) {
+        if (!is_array($spotMsgIdList) || count($spotMsgIdList) == 0) {
             return $tmparray;
         } // if
 
         // prepare a list of IN values
         $msgIdList = $this->_conn->arrayKeyToIn($spotMsgIdList);
+
+        if (!isset($msgIdList) || $msgIdList == '') {
+            return $tmparray;
+        } // if
+
+        SpotTiming::start(__CLASS__.'::'.__FUNCTION__);
+
         $msgIdList = '('.$msgIdList.')';
 
         $tmpArray = $this->_conn->arrayQuery('SELECT s.messageid AS messageid, s.spotterid AS spotterid, s.stamp AS stamp
@@ -372,14 +400,18 @@ class Dao_Base_Spot implements Dao_Spot
     public function removeSpots($spotMsgIdList)
     {
         // Empty list provided? Exit
-        if (count($spotMsgIdList) == 0) {
+        if (!is_array($spotMsgIdList) || count($spotMsgIdList) == 0) {
+            return;
+        } // if
+
+        // prepare a list of IN values
+        $msgIdList = $this->_conn->arrayKeyToIn($spotMsgIdList);
+
+        if (!isset($msgIdList) || $msgIdList == '') {
             return;
         } // if
 
         SpotTiming::start(__CLASS__.'::'.__FUNCTION__);
-
-        // prepare a list of IN values
-        $msgIdList = $this->_conn->arrayKeyToIn($spotMsgIdList);
 
         $this->_conn->modify('DELETE FROM spots WHERE messageid IN ('.$msgIdList.')');
         $this->_conn->modify('DELETE FROM spotsfull WHERE messageid  IN ('.$msgIdList.')');
@@ -400,14 +432,20 @@ class Dao_Base_Spot implements Dao_Spot
     public function markSpotsModerated($spotMsgIdList)
     {
         // Empty list provided? Exit
-        if (count($spotMsgIdList) == 0) {
+        if (!is_array($spotMsgIdList) || count($spotMsgIdList) == 0) {
+            return;
+        } // if
+
+        // prepare a list of IN values
+        $msgIdList = $this->_conn->arrayKeyToIn($spotMsgIdList);
+
+        if (!isset($msgIdList) || $msgIdList == '') {
             return;
         } // if
 
         SpotTiming::start(__CLASS__.'::'.__FUNCTION__);
         $this->_conn->modify(
-            'UPDATE spots SET moderated = :moderated WHERE messageid IN ('.
-                                $this->_conn->arrayKeyToIn($spotMsgIdList).')',
+            'UPDATE spots SET moderated = :moderated WHERE messageid IN ('.$msgIdList.')',
             [
                 ':moderated' => [true, PDO::PARAM_BOOL],
             ]
@@ -658,14 +696,18 @@ class Dao_Base_Spot implements Dao_Spot
         $idList = ['spot' => [], 'fullspot' => []];
 
         // Empty list, exit
-        if (count($hdrList) == 0) {
+        if (!is_array($hdrList) || count($hdrList) == 0) {
+            return $idList;
+        } // if
+
+        // Prepare a list of values
+        $msgIdList = $this->_conn->arrayValToIn($hdrList, 'Message-ID');
+
+        if (!isset($msgIdList) || $msgIdList == '') {
             return $idList;
         } // if
 
         SpotTiming::start(__CLASS__.'::'.__FUNCTION__);
-
-        // Prepare a list of values
-        $msgIdList = $this->_conn->arrayValToIn($hdrList, 'Message-ID');
 
         // Because MySQL doesn't know anything about full joins, we use this trick
         $rs = $this->_conn->arrayQuery("SELECT messageid AS spot, '' AS fullspot FROM spots WHERE messageid IN (".$msgIdList.")
